feat: state events

+++ src/Gang.Tests/FakeGangMember.cs
@@ -38,7 +38,7 @@ namespace Gang.Tests
         }
 
         Task IGangMember.SendAsync(GangMessageTypes type,
-            byte[] data, byte[] memberId, short? sequenceNumber)
+            byte[] data, byte[] memberId, uint? sequenceNumber)
         {
             MessagesReceived.Add(new Message(type, data, sequenceNumber));
             return Task.CompletedTask;
@@ -54,7 +54,7 @@ namespace Gang.Tests
         public class Message
         {
             public Message(
-                GangMessageTypes type, byte[] data, short? sequenceNumber)
+                GangMessageTypes type, byte[] data, uint? sequenceNumber)
             {
                 Type = type;
                 Data = data;
@@ -63,7 +63,7 @@ namespace Gang.Tests
 
             public GangMessageTypes Type { get; }
             public byte[] Data { get; }
-            public short? SequenceNumber { get; }
+            public uint? SequenceNumber { get; }
         }
     }
 }
